Title: 실수 받아 그대로 출력, Time: 145ms, Memory: 12MB, Status: Passed - CodeTree

+++ 241021/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double n = sc.nextInt();
+        double n = sc.nextDouble();
 
         System.out.printf("%.2f",n);
     }
